moved all get routes for curr user to appropraite route files to match postman test routes
