+++ backend/routes/api/spots.js
@@ -112,6 +112,57 @@ router.get('/', async (req, res) => {
 
 });
 
+//~GET SPOTS OWNED BY CURRENT USER
+router.get('/current', requireAuth, async (req, res) => {
+  try {
+      const spots = await Spot.findAll({
+        where: {
+          ownerId: req.user.id
+        },
+        include: [{
+          model: Image,
+          as: 'Images',
+          where: { preview: true },
+          attributes: ['url'],
+          limit: 1
+        }],
+      });
+
+      const reviewAverages = await Review.findAll({
+          attributes: ['spotId', [fn('SUM', col('stars')), 'sumStars'], [fn('COUNT', col('stars')), 'reviewCount']],
+          group: ['spotId']
+      });
+
+      const avgRateMap = {};
+      reviewAverages.forEach( reviewAverage => {
+          const { spotId, sumStars, reviewCount } = reviewAverage.dataValues;
+          const avgRating = sumStars / reviewCount;
+          avgRateMap[spotId] = avgRating;
+      });
+
+      const responseData = spots.map(spot => {
+          const spotObj = spot.get();
+
+          delete spotObj.Images;
+
+          return {
+              ...spotObj,
+              createdAt: formatDate(spot.createdAt),
+              updatedAt: formatDate(spot.updatedAt),
+              avgRating: avgRateMap[spot.id] || 1,
+              previewImage: spot.Images[0]?.url || null
+          }
+      });
+
+      res.json({ Spots: responseData });
+  } catch (error) {
+    console.error('Error fetching spot:', error);
+    return res.status(500).json({ error: 'Internal Server Error' });
+  };
+
+
+});
+
 
 //~ GET SPOT DETAILS BY SPOT ID
 router.get('/:spotId', async (req, res) => {
